% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httppipe.R
\name{httppipe}
\alias{httppipe}
\title{Create a httppipe connector}
\usage{
httppipe(path)
}
\arguments{
\item{path}{The path of the pipe or socket}
}
\description{
Construct a httppipe connector.  This returns a function that can
be used to make requests over a unix socket or named pipe
(depending on the platform).
}
