% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httppipe.R
\name{httppipe_available}
\alias{httppipe_available}
\title{Test if httppipe available}
\usage{
httppipe_available(verbose = FALSE)
}
\arguments{
\item{verbose}{Print information about failure}
}
\description{
Test httppipe support is available.  This tests that the python
module can be loaded (not that any actual server is running on a
socket or named pipe).  This is designed to be used in packages
that depend on this package as an efficient way of testing if it
can be used.
}
